+++ README.md
@@ -2,8 +2,25 @@
 
 This is the code for the Sigma Phi Delta Fraternity website.
 
-### To run: ###
+### To install: ###
+Note: this assumes you have a working Node.js and MongoDB setup.
+Run the following command to install dependencies:
+```
+npm install
+```
 
-```javascript
-npm start
+### To run: ###
+1) Start the MongoDB server
+```
+mongod --dbpath ./db/data
+```
+
+2) Start the HTTP redirect server
+```
+npm run-script http
+```
+
+3) Start the HTTPS server
+```
+npm run-script https
 ```
