+++ app.js
@@ -1,39 +1,56 @@
-// Imports
-var express = require('express');
-var cookieParser = require('cookie-parser');
-var bodyParser = require('body-parser');
-var favicon = require("serve-favicon");
-
-var route_controller = require('./controllers/routes');
-var handler_controller = require("./controllers/handlers");
-
-var app = express();
-
-// Templating engine
-app.set("views", __dirname + "/views");
-app.set("view engine", "ejs");
-
-// Use frameworks
-app.use(favicon(__dirname + "/public/images/favicon.ico"));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cookieParser());
-app.use(express.static(__dirname + "/public"));
-
-// Find route
-app.use("/", route_controller);
-app.use("/", handler_controller);
-
-// catch 404
-app.use(function(request, response) {
-  var err = new Error("Page not found.");
-  response.status(404);
-  response.render("template", {
-    title: "Not Found",
-    user: null,
-    message: err.message,
-    error: err
+var createApplication = function() {
+  
+  // Middleware imports
+  var express = require("express");
+  var cookieParser = require("cookie-parser");
+  var bodyParser = require("body-parser");
+  var favicon = require("serve-favicon");
+  var mongoose = require("mongoose");
+  
+  // Local imports
+  var sessionManager = require(__dirname + "/util/SessionManager")();
+  var sanitizer = require(__dirname + "/util/Sanitizer")();
+  var renderController = require(__dirname + "/controllers/render");
+  var apiController = require(__dirname + "/controllers/api");
+  var userSchema = require(__dirname + "/db/models/user")(mongoose.Schema);
+  
+  // Create application
+  var app = express();
+  
+  // Templating engine
+  app.set("views", __dirname + "/views");
+  app.set("view engine", "ejs");
+  
+  // Middleware
+  app.use(favicon(__dirname + "/public/images/favicon.ico"));
+  app.use(bodyParser.json());
+  app.use(bodyParser.urlencoded({ extended: false }));
+  app.use(cookieParser());
+  
+  // Create context for url mapping controllers
+  var context = {
+    "models": {
+      "User": mongoose.model("User", userSchema)
+    },
+    "sessionManager": sessionManager,
+    "sanitizer": sanitizer
+  };
+  
+  // Serve mapped urls
+  app.use("/", renderController(express.Router(), context));
+  app.use("/", apiController(express.Router(), context));
+  app.use(express.static(__dirname + "/public"));
+  
+  // Serve unmapped urls
+  app.use(function(request, response) {
+    response.render("error", {
+      title: "error",
+      user: null,
+    });
   });
-});
+  
+  return app;
+  
+};
 
-module.exports = app;
+module.exports = createApplication;
