--- bin/server/https.js
@@ -1,13 +1,0 @@
-var app = require("../../app");
-var https = require("https");
-var fs = require("fs");
-
-var ssl_options = {
-  key: fs.readFileSync(__dirname + "/../keys/ssl.key"),
-  cert: fs.readFileSync(__dirname + "/../keys/ssl.crt"),
-  ca: fs.readFileSync(__dirname + "/../keys/sub.class1.server.ca.pem")
-};
-
-https.createServer(ssl_options, app).listen(9001);
-
-console.log("https server started.");
