+++ package.json
@@ -3,18 +3,16 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "http": "node ./bin/server/http.js >> ./bin/logs/http.log",
-    "https": "node ./bin/server/https.js >> ./bin/logs/https.log",
-    "test": "node ./bin/server/test.js >> ./bin/logs/test.log"
+    "http": "node ./server/http.js",
+    "https": "node ./server/https.js",
+    "test": "node ./server/test.js"
   },
   "dependencies": {
     "body-parser": "~1.12.4",
     "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
     "ejs": "~2.3.1",
     "express": "~4.12.4",
-    "serve-favicon": "~2.2.1",
-    "aws-sdk": "~2.1.24",
-    "minify": "~1.4.19"
+    "mongoose": "^4.2.4",
+    "serve-favicon": "~2.2.1"
   }
 }
