+++ server/http.js
@@ -3,10 +3,8 @@ var http = require("http");
 http.createServer(function(request, response) {
   
   response.writeHead(301, {
-    "Location": "https://spdcalpoly.org" + request.url
+    "Location": "https://" + request.headers.host + request.url
   });
   response.end();
 
-}).listen(9000);
-
-console.log("Http server started.");
+}).listen(9000 || process.argv[2]);
